moved node_info
