+++ src/accuracy_checker/executors.py
@@ -104,7 +104,7 @@ class docker_executor(executor):
         return out
 
     def get_infrastructure(self):
-        hardware_command = 'python3 /tmp/dl-benchmark/src/accuracy_checker/node_info.py'
+        hardware_command = 'python3 /tmp/dl-benchmark/src/node_info/node_info.py'
         command_line = 'bash -c "source /root/.bashrc && {}"'.format(hardware_command)
         output = self.my_container_dict[self.my_target_framework].exec_run(command_line, tty=True, privileged=True)
         if output[0] != 0:
