+++ src/benchmark/executors.py
@@ -41,7 +41,7 @@ class host_executor(executor):
         return '../inference'
 
     def get_infrastructure(self):
-        import node_info as info
+        from src.node_info import node_info as info
         hardware = info.get_system_characteristics()
         hardware_info = ''
         for key in hardware:
@@ -67,7 +67,7 @@ class docker_executor(executor):
         return '/tmp/dl-benchmark/src/inference'
 
     def get_infrastructure(self):
-        hardware_command = 'python3 /tmp/dl-benchmark/src/benchmark/node_info.py'
+        hardware_command = 'python3 /tmp/dl-benchmark/src/node_info/node_info.py'
         command_line = 'bash -c "source /root/.bashrc && {}"'.format(hardware_command)
         output = self.my_container_dict[self.my_target_framework].exec_run(command_line, tty=True, privileged=True)
         if output[0] != 0:
